fix: eslint errors and type errors
