+++ frontend/src/routes/problems/$problemId/-contexts/problem-context/problem-context.tsx
@@ -6,7 +6,6 @@ import {
   FC,
   ReactNode,
   SetStateAction,
-  useContext,
   useState,
 } from "react"
 
@@ -34,11 +33,15 @@ export const ProblemProvider: FC<ProblemProviderProps> = ({
   children,
   problemId,
 }) => {
-  const { data: problem } = $api.useSuspenseQuery("get", "/api/problems/{id}", {
-    params: {
-      path: { id: problemId },
+  const { data: problem } = $api.useSuspenseQuery(
+    "get",
+    "/api/problems/{problemId}",
+    {
+      params: {
+        path: { problemId },
+      },
     },
-  })
+  )
 
   const [language, setLanguage] = useState<Language>(
     problem.supported_languages[0],
@@ -52,12 +55,4 @@ export const ProblemProvider: FC<ProblemProviderProps> = ({
       {children}
     </ProblemContext.Provider>
   )
-}
-
-export const useProblem = () => {
-  const context = useContext(ProblemContext)
-  if (context === undefined) {
-    throw new Error("useProblem must be used within a ProblemProvider")
-  }
-  return context
 }
