+++ frontend/src/routes/problems/$problemId/-hooks/use-problem/index.ts
@@ -1,0 +1,1 @@
+export { useProblem } from "./use-problem"
